+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatComboBoxUI.java
@@ -480,13 +480,8 @@ public class FlatComboBoxUI
 
 		@Override
 		protected Rectangle computePopupBounds( int px, int py, int pw, int ph ) {
-			// get maximum display size of all items, ignoring prototype value
-			Object prototype = comboBox.getPrototypeDisplayValue();
-			if( prototype != null )
-				comboBox.setPrototypeDisplayValue( null );
+			// get maximum display size of all items
 			Dimension displaySize = getDisplaySize();
-			if( prototype != null )
-				comboBox.setPrototypeDisplayValue( prototype );
 
 			// make popup wider if necessary
 			if( displaySize.width > pw ) {