+++ flatlaf-demo/src/main/java/com/formdev/flatlaf/demo/BasicComponentsPanel.java
@@ -310,14 +310,13 @@ class BasicComponentsPanel
 		add(comboBox4, "cell 4 4,growx");
 
 		//---- comboBox5 ----
-		comboBox5.setPrototypeDisplayValue("12345");
 		comboBox5.setModel(new DefaultComboBoxModel<>(new String[] {
 			"wide popup if text is longer",
 			"aa",
 			"bbb",
 			"cccc"
 		}));
-		add(comboBox5, "cell 5 4,growx");
+		add(comboBox5, "cell 5 4,growx,wmax 100");
 
 		//---- spinnerLabel ----
 		spinnerLabel.setText("JSpinner:");